changed styles SortFilter, delete ProductSection,add cursor pointer to btn in Filter
