--- src/components/ProductsSection/productsSection.scss
@@ -1,3 +1,0 @@
-.products-section {
-   @include margin(63px, 95px);
-}
