+++ src/components/SortFilter/sortFilter.scss
@@ -1,10 +1,13 @@
 .filter-section-sort{
     position: absolute;
-    right: 20px;
-    top: 200px;
+    top: 50px;
+    &--position {
+        right: 20px;
+        top: -5px;
+    }
 
     &-select {
-       width: 113px;
+    width: 113px;
     height: 24px;
     border: 1px solid $secondary;
     border-radius: 10px;
