+++ src/pages/Products/products.scss
@@ -1,0 +1,4 @@
+.products-section {
+    position: relative;
+    @include margin(33px, 95px);
+ }
