+++ src/styles/style.scss
@@ -11,10 +11,10 @@
 @import "../pages/NotFound/notFound";
 @import "../pages/About/about";
 @import "../pages/CartPage/cartPage";
+@import "../pages/Products/products";
 
 
 //!Components
-@import "../components/ProductsSection/productsSection";
 @import "../components/ProductCard/productCard";
 @import "../components/OurBenefitsList/ourBenefits";
 @import "../components/OurPartners/ourPartners";
